% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA_Stack.R
\name{as.data.frame.dta_stack}
\alias{as.data.frame.dta_stack}
\title{Coerce DTA Stack Object to Data Frame}
\usage{
\method{as.data.frame}{dta_stack}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{dta_stack object}

\item{row.names}{NULL or a character vector}

\item{optional}{logical}

\item{...}{Additional arguments}
}
\description{
Coerce DTA Stack Object to Data Frame
}
