% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{AuditC_data}
\alias{AuditC_data}
\title{Alcohol Screening Test Accuracy (AUDIT-C)}
\format{
A data frame with diagnostic accuracy data including:
\describe{
  \item{TP}{True positives}
  \item{FP}{False positives}
  \item{FN}{False negatives}
  \item{TN}{True negatives}
}
}
\source{
Sourced from the \code{mada} package. Original data from systematic
reviews of alcohol screening test accuracy.
}
\usage{
AuditC_data
}
\description{
Diagnostic test accuracy data for the AUDIT-C (Alcohol Use Disorders
Identification Test - Consumption) screening tool for detecting unhealthy
alcohol use.
}
\examples{
data(AuditC_data)
head(AuditC_data)
# Calculate sensitivity and specificity
AuditC_data$sensitivity <- with(AuditC_data, TP/(TP+FN))
AuditC_data$specificity <- with(AuditC_data, TN/(TN+FP))
}
\references{
Doebler P, Holling H (2015). Meta-analysis of Diagnostic Accuracy with mada.
\url{https://CRAN.R-project.org/package=mada}
}
\keyword{datasets}
