% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{COVID_AntigenTests_Cochrane2021}
\alias{COVID_AntigenTests_Cochrane2021}
\title{COVID-19 Rapid Antigen Tests - Cochrane Review (2021)}
\format{
A data frame with 20 studies:
\describe{
  \item{author}{First author name}
  \item{year}{Publication year}
  \item{TP}{True positives}
  \item{FP}{False positives}
  \item{FN}{False negatives}
  \item{TN}{True negatives}
  \item{test_brand}{Commercial test brand name}
  \item{setting}{Testing setting (community or hospital)}
}
}
\source{
Based on Cochrane review by Dinnes J, et al. (2021).
}
\usage{
COVID_AntigenTests_Cochrane2021
}
\description{
Diagnostic accuracy of rapid antigen tests for SARS-CoV-2 infection.
From a Cochrane systematic review evaluating point-of-care tests across
different settings and test brands.
}
\examples{
data(COVID_AntigenTests_Cochrane2021)
head(COVID_AntigenTests_Cochrane2021)
# Sensitivity by setting
tapply(with(COVID_AntigenTests_Cochrane2021, TP/(TP+FN)),
       COVID_AntigenTests_Cochrane2021$setting, mean)
}
\references{
Dinnes J, Deeks JJ, Berhane S, et al. (2021). Rapid, point-of-care antigen
and molecular-based tests for diagnosis of SARS-CoV-2 infection.
\emph{Cochrane Database of Systematic Reviews}, \bold{3}, CD013705.
}
\keyword{datasets}
