% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{create_dta_report}
\alias{create_dta_report}
\title{Create DTA-Stack TruthCert Report}
\usage{
create_dta_report(x, filename = "DTA_TruthCert_Report.html")
}
\arguments{
\item{x}{dta_stack object}

\item{filename}{Output filename}
}
\description{
Create DTA-Stack TruthCert Report
}
