% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{DDimer_PE_Crawford2020}
\alias{DDimer_PE_Crawford2020}
\title{D-Dimer for Pulmonary Embolism - Crawford et al. (2020)}
\format{
A data frame with 22 studies:
\describe{
  \item{author}{First author name}
  \item{year}{Publication year}
  \item{TP}{True positives}
  \item{FP}{False positives}
  \item{FN}{False negatives}
  \item{TN}{True negatives}
  \item{d_dimer_threshold}{D-dimer cutoff threshold (ng/mL)}
  \item{setting}{Clinical setting (outpatient, ED, hospital)}
}
}
\source{
Based on Crawford F, et al. (2020).
}
\usage{
DDimer_PE_Crawford2020
}
\description{
Diagnostic accuracy of D-dimer testing for suspected pulmonary embolism.
From a comprehensive systematic review published in Blood Advances in 2020.
Includes studies with various D-dimer thresholds.
}
\examples{
data(DDimer_PE_Crawford2020)
head(DDimer_PE_Crawford2020)
# Sensitivity and specificity
DDimer_PE_Crawford2020$sensitivity <- with(DDimer_PE_Crawford2020, TP/(TP+FN))
DDimer_PE_Crawford2020$specificity <- with(DDimer_PE_Crawford2020, TN/(TN+FP))
summary(DDimer_PE_Crawford2020[,c("sensitivity", "specificity")])
}
\references{
Crawford F, Andras A, Welch K, et al. (2020). Systematic review and
meta-analysis of test accuracy for the diagnosis of suspected pulmonary
embolism. \emph{Blood Advances}, \bold{4}(18), 4296-4311.
}
\keyword{datasets}
