% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{DVT_Ultrasound_Goodacre2005}
\alias{DVT_Ultrasound_Goodacre2005}
\title{Ultrasound for Deep Vein Thrombosis - Goodacre et al. (2005)}
\format{
A data frame with 15 studies:
\describe{
  \item{author}{First author name}
  \item{year}{Publication year}
  \item{TP}{True positives}
  \item{FP}{False positives}
  \item{FN}{False negatives}
  \item{TN}{True negatives}
  \item{ultrasound_type}{Type of ultrasound (CUS or Duplex)}
  \item{population}{Patient population}
}
}
\source{
Based on Goodacre S, et al. (2005).
}
\usage{
DVT_Ultrasound_Goodacre2005
}
\description{
Diagnostic accuracy of ultrasound for diagnosing deep vein thrombosis (DVT).
Includes both compression ultrasound (CUS) and duplex ultrasound methods.
}
\examples{
data(DVT_Ultrasound_Goodacre2005)
head(DVT_Ultrasound_Goodacre2005)
# Compare ultrasound types
table(DVT_Ultrasound_Goodacre2005$ultrasound_type)
}
\references{
Goodacre S, Sampson F, Thomas S, van Beek E, Sutton A (2005). Systematic
review and meta-analysis of the diagnostic accuracy of ultrasonography for
deep vein thrombosis. \emph{BMC Medical Imaging}, \bold{5}, 6.
}
\keyword{datasets}
