% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{IAQ_data}
\alias{IAQ_data}
\title{IAQ (Information/Annoyance Questionnaire) Data}
\format{
A data frame with diagnostic accuracy data including:
\describe{
  \item{author}{First author name}
  \item{study_id}{Study identifier}
  \item{result_id}{Result identifier}
  \item{type}{Type of study or test}
  \item{TP}{True positives}
  \item{FN}{False negatives}
  \item{FP}{False positives}
  \item{TN}{True negatives}
  \item{population}{Population studied}
}
}
\source{
Sourced from the \code{mada} package.
}
\usage{
IAQ_data
}
\description{
Diagnostic test accuracy data for the IAQ screening instrument.
}
\examples{
data(IAQ_data)
head(IAQ_data)
}
\references{
Doebler P, Holling H (2015). Meta-analysis of Diagnostic Accuracy with mada.
\url{https://CRAN.R-project.org/package=mada}
}
\keyword{datasets}
