% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{MRI_Prostate_Futterer2015}
\alias{MRI_Prostate_Futterer2015}
\title{MRI for Prostate Cancer Detection - Futterer et al. (2015)}
\format{
A data frame with 15 studies:
\describe{
  \item{author}{First author name}
  \item{year}{Publication year}
  \item{TP}{True positives}
  \item{FP}{False positives}
  \item{FN}{False negatives}
  \item{TN}{True negatives}
  \item{mri_strength}{MRI field strength (1.5T or 3.0T)}
  \item{pirads_score}{PI-RADS score threshold (3, 4, or 5)}
}
}
\source{
Based on published meta-analyses of mpMRI for prostate cancer.
}
\usage{
MRI_Prostate_Futterer2015
}
\description{
Diagnostic accuracy of multiparametric MRI for detecting prostate cancer.
Includes data on different MRI field strengths and PI-RADS scores.
}
\examples{
data(MRI_Prostate_Futterer2015)
head(MRI_Prostate_Futterer2015)
# Compare by MRI strength
table(MRI_Prostate_Futterer2015$mri_strength)
}
\references{
Futterer JJ, Briganti A, De Visschere P, et al. (2015). Can clinically
significant prostate cancer be detected with multiparametric magnetic
resonance imaging? A systematic review of the literature.
\emph{European Urology}, \bold{68}(6), 1045-1053.
}
\keyword{datasets}
