% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA_Stack.R
\name{plot.dta_stack}
\alias{plot.dta_stack}
\title{Plot DTA Stack Object}
\usage{
\method{plot}{dta_stack}(x, type = "sroc", ...)
}
\arguments{
\item{x}{dta_stack object}

\item{type}{Plot type ("sroc", "entropy", "subgroups", or "fagan")}

\item{...}{Additional arguments}
}
\description{
Plot DTA Stack Object
}
