% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{POCUS_Shock_Yoshida2023}
\alias{POCUS_Shock_Yoshida2023}
\title{Point-of-Care Ultrasound for Shock - Yoshida et al. (2023)}
\format{
A data frame with 12 studies:
\describe{
  \item{author}{First author name}
  \item{year}{Publication year}
  \item{TP}{True positives}
  \item{FP}{False positives}
  \item{FN}{False negatives}
  \item{TN}{True negatives}
  \item{shock_type}{Type of shock (cardiogenic, distributive, hypovolemic, obstructive, mixed)}
  \item{country}{Country where study was conducted}
}
}
\source{
Based on Yoshida T, et al. (2023).
}
\usage{
POCUS_Shock_Yoshida2023
}
\description{
Diagnostic accuracy of point-of-care ultrasound (POCUS) for identifying
the cause of shock in critically ill patients. From a 2023 systematic
review published in Critical Care.
}
\examples{
data(POCUS_Shock_Yoshida2023)
head(POCUS_Shock_Yoshida2023)
# Sensitivity by shock type
POCUS_Shock_Yoshida2023$sensitivity <- with(POCUS_Shock_Yoshida2023, TP/(TP+FN))
tapply(POCUS_Shock_Yoshida2023$sensitivity,
       POCUS_Shock_Yoshida2023$shock_type, mean)
}
\references{
Yoshida T, Endo A, Oiwa A, et al. (2023). Diagnostic accuracy of point-of-care
ultrasound for shock: a systematic review and meta-analysis.
\emph{Critical Care}, \bold{27}(1), 200.
}
\keyword{datasets}
