% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA_Stack.R
\name{print.dta_stack}
\alias{print.dta_stack}
\title{Print DTA Stack Object}
\usage{
\method{print}{dta_stack}(x, ...)
}
\arguments{
\item{x}{dta_stack object}

\item{...}{Additional arguments}
}
\description{
Print DTA Stack Object
}
