% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Procalcitonin_Sepsis_Wacker2013}
\alias{Procalcitonin_Sepsis_Wacker2013}
\title{Procalcitonin for Sepsis Diagnosis - Wacker et al. (2013)}
\format{
A data frame with 18 studies:
\describe{
  \item{author}{First author name}
  \item{year}{Publication year}
  \item{TP}{True positives}
  \item{FP}{False positives}
  \item{FN}{False negatives}
  \item{TN}{True negatives}
  \item{cutoff_ng_ml}{Procalcitonin cutoff threshold (ng/mL)}
  \item{setting}{Clinical setting (ICU, ED, hospital)}
}
}
\source{
Based on Wacker C, et al. (2013).
}
\usage{
Procalcitonin_Sepsis_Wacker2013
}
\description{
Diagnostic accuracy of procalcitonin as a marker for sepsis in critically
ill patients. From a Lancet Infectious Diseases meta-analysis with various
cutoff values and clinical settings.
}
\examples{
data(Procalcitonin_Sepsis_Wacker2013)
head(Procalcitonin_Sepsis_Wacker2013)
# Distribution of cutoffs
hist(Procalcitonin_Sepsis_Wacker2013$cutoff_ng_ml,
     main="Procalcitonin Cutoff Values", xlab="Cutoff (ng/mL)")
}
\references{
Wacker C, Prkno A, Brunkhorst FM, Schlattmann P (2013). Procalcitonin as a
diagnostic marker for sepsis: a systematic review and meta-analysis.
\emph{Lancet Infectious Diseases}, \bold{13}(5), 426-435.
}
\keyword{datasets}
