% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA_Stack.R
\name{run_bayesian_mcmc}
\alias{run_bayesian_mcmc}
\title{Run Bayesian MCMC Meta-Analysis with R-hat diagnostics}
\usage{
run_bayesian_mcmc(d, iter)
}
\description{
Run Bayesian MCMC Meta-Analysis with R-hat diagnostics
}
\keyword{internal}
