% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA_Stack.R
\name{run_firth_penalized}
\alias{run_firth_penalized}
\title{Run Penalized Likelihood (Firth) for sparse data}
\usage{
run_firth_penalized(d)
}
\description{
Run Penalized Likelihood (Firth) for sparse data
}
\keyword{internal}
