% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA_Stack.R
\name{run_model_averaging}
\alias{run_model_averaging}
\title{Dynamic Model Averaging (Bayesian + Frequentist)}
\usage{
run_model_averaging(d, iter)
}
\description{
Dynamic Model Averaging (Bayesian + Frequentist)
}
\keyword{internal}
