% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_logic.R
\name{run_regression_glmm}
\alias{run_regression_glmm}
\title{Run Regression GLMM (Bivariate Logit with Covariates)}
\usage{
run_regression_glmm(data, mods)
}
\description{
Run Regression GLMM (Bivariate Logit with Covariates)
}
\keyword{internal}
