% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SkinTests_data}
\alias{SkinTests_data}
\title{Skin Tests Diagnostic Accuracy Data}
\format{
A data frame with diagnostic accuracy data including:
\describe{
  \item{TP}{True positives}
  \item{FP}{False positives}
  \item{FN}{False negatives}
  \item{TN}{True negatives}
}
}
\source{
Sourced from the \code{mada} package.
}
\usage{
SkinTests_data
}
\description{
Diagnostic test accuracy data from studies evaluating skin tests for various
conditions.
}
\examples{
data(SkinTests_data)
head(SkinTests_data)
}
\references{
Doebler P, Holling H (2015). Meta-analysis of Diagnostic Accuracy with mada.
\url{https://CRAN.R-project.org/package=mada}
}
\keyword{datasets}
