% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA_Stack.R
\name{summary.dta_stack}
\alias{summary.dta_stack}
\title{Summary of DTA Stack Object}
\usage{
\method{summary}{dta_stack}(object, ...)
}
\arguments{
\item{object}{dta_stack object}

\item{...}{Additional arguments}
}
\description{
Summary of DTA Stack Object
}
