% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{TB_SmearMicroscopy_Steingart2006}
\alias{TB_SmearMicroscopy_Steingart2006}
\title{Sputum Smear Microscopy for Tuberculosis - Steingart et al. (2006)}
\format{
A data frame with 20 studies:
\describe{
  \item{author}{First author name}
  \item{year}{Publication year}
  \item{TP}{True positives}
  \item{FP}{False positives}
  \item{FN}{False negatives}
  \item{TN}{True negatives}
  \item{smear_method}{Microscopy staining method (ZN, Auramine, LED)}
  \item{country}{Country where study was conducted}
}
}
\source{
Based on Steingart KR, et al. (2006).
}
\usage{
TB_SmearMicroscopy_Steingart2006
}
\description{
Diagnostic accuracy of sputum smear microscopy for diagnosing pulmonary
tuberculosis. Includes data on different staining methods (Ziehl-Neelsen,
Auramine, LED) across multiple countries.
}
\examples{
data(TB_SmearMicroscopy_Steingart2006)
head(TB_SmearMicroscopy_Steingart2006)
# Compare methods
table(TB_SmearMicroscopy_Steingart2006$smear_method)
}
\references{
Steingart KR, Henry M, Ng V, et al. (2006). Fluorescence versus conventional
sputum smear microscopy for tuberculosis: a systematic review.
\emph{Lancet Infectious Diseases}, \bold{6}(9), 570-581.
}
\keyword{datasets}
