% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{XpertMTB_RIF_Tuberculosis2014}
\alias{XpertMTB_RIF_Tuberculosis2014}
\title{Xpert MTB/RIF for Tuberculosis - Steingart et al. (2014)}
\format{
A data frame with 17 studies:
\describe{
  \item{author}{First author name}
  \item{year}{Publication year}
  \item{TP}{True positives}
  \item{FP}{False positives}
  \item{FN}{False negatives}
  \item{TN}{True negatives}
  \item{sample_type}{Type of sample (sputum, various)}
  \item{hiv_status}{HIV status of population (positive, negative, mixed)}
  \item{country}{Country where study was conducted}
}
}
\source{
Based on Steingart KR, et al. (2014).
}
\usage{
XpertMTB_RIF_Tuberculosis2014
}
\description{
Diagnostic accuracy of Xpert MTB/RIF for detecting pulmonary tuberculosis
and rifampicin resistance. From a Cochrane systematic review including
studies from various countries with different HIV prevalence.
}
\examples{
data(XpertMTB_RIF_Tuberculosis2014)
head(XpertMTB_RIF_Tuberculosis2014)
# Accuracy by HIV status
XpertMTB_RIF_Tuberculosis2014$sensitivity <- with(XpertMTB_RIF_Tuberculosis2014,
                                                    TP/(TP+FN))
tapply(XpertMTB_RIF_Tuberculosis2014$sensitivity,
       XpertMTB_RIF_Tuberculosis2014$hiv_status, mean)
}
\references{
Steingart KR, Schiller I, Horne DJ, et al. (2014). Xpert MTB/RIF assay for
pulmonary tuberculosis and rifampicin resistance in adults. \emph{Cochrane
Database of Systematic Reviews}, \bold{1}, CD009593.
}
\keyword{datasets}
